transboltBackend/requirements.txt
asgiref==3.7.2
bson==0.5.10
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1